Abstract content and search screens. Unifies gesture navigation
